Campagne sonar

+++ batimen-web/src/main/java/fr/batimen/web/client/extend/connected/Annonce.java
@@ -329,8 +329,8 @@ public class Annonce extends MasterPage {
 
                 Label categorie = new Label("categorie", Categorie.getNameByCode(item.getModelObject()));
 
-                if (item.getModelObject().equals(Categorie.PLOMBERIE_CODE.shortValue())
-                        || item.getModelObject().equals(Categorie.ESPACE_VERT_CODE.shortValue())) {
+                if (item.getModelObject().equals(Categorie.PLOMBERIE_CODE)
+                        || item.getModelObject().equals(Categorie.ESPACE_VERT_CODE)) {
                     categorie.add(new AttributeModifier("class", "labelAnnonce-icon8"));
                 }
 
@@ -451,6 +451,7 @@ public class Annonce extends MasterPage {
 
                     @Override
                     public void onClick() {
+                        //TODO : A implementer
                     }
                 };
